+++ randomContact/src/main/java/com/ae/apps/randomcontact/permissions/ContactsPermissionFragment.kt
@@ -1,23 +1,33 @@
+/*
+ * Copyright 2018 Midhun Harikumar
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.ae.apps.randomcontact.permissions
 
 import android.Manifest
 import android.view.View
 import android.widget.Button
 import android.widget.TextView
-
 import com.ae.apps.randomcontact.R
-import com.ae.apps.randomcontact.utils.AppConstants
 
 /**
  * An abstract implementation of PermissionCheckingFragment for Read Contact permission
  */
 abstract class ContactsPermissionFragment : PermissionCheckingFragment() {
-    override fun getRequestCode(): Int {
-        return AppConstants.PERMISSIONS_REQUEST_READ_CONTACTS
-    }
 
     override fun getRequiredPermissions(): Array<String> {
-        return arrayOf(Manifest.permission.READ_CONTACTS)
+        return arrayOf(Manifest.permission.READ_CONTACTS, Manifest.permission.CALL_PHONE)
     }
 
     public override fun setupViewWithoutPermission(): View {