+++ randomContact/src/main/java/com/ae/apps/randomcontact/fragments/RandomContactFragment.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015 Midhun Harikumar
+ * Copyright 2015-2018 Midhun Harikumar
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,22 +16,20 @@
 
 package com.ae.apps.randomcontact.fragments;
 
-import android.content.Context;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import android.support.annotation.NonNull;
 import android.support.v7.graphics.Palette;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
-import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
-import android.view.ViewGroup;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
+import android.widget.Button;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 import android.widget.Toast;
@@ -39,16 +37,21 @@ import android.widget.Toast;
 import com.ae.apps.common.managers.contact.AeContactManager;
 import com.ae.apps.common.views.RoundedImageView;
 import com.ae.apps.common.vo.ContactVo;
+import com.ae.apps.common.vo.PhoneNumberVo;
 import com.ae.apps.randomcontact.R;
 import com.ae.apps.randomcontact.adapters.ContactRecyclerAdapter;
 import com.ae.apps.randomcontact.data.GlobalThemeChanger;
 import com.ae.apps.randomcontact.managers.RandomContactManager;
+import com.ae.apps.randomcontact.permissions.ContactsPermissionFragment;
 import com.ae.apps.randomcontact.utils.AppConstants;
 import com.ae.apps.randomcontact.utils.Utils;
 
 import java.util.Collections;
 
-public class RandomContactFragment extends Fragment {
+/**
+ * Fragment that shows a Random Contact
+ */
+public class RandomContactFragment extends ContactsPermissionFragment {
 
     private static final String SAVED_CONTACT_ID = "savedContactId";
 
@@ -67,29 +70,46 @@ public class RandomContactFragment extends Fragment {
     private Bitmap mDefaultUserImage;
     private ContactRecyclerAdapter mRecyclerAdapter;
     private GlobalThemeChanger mContactManagerProvider;
-    private Context mContext;
     private AeContactManager mContactManager;
     private ContactVo mCurrentContact;
 
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View layout = inflater.inflate(R.layout.fragment_random_contact, container, false);
+    private View setupViewWithoutContacts() {
+        View noAccessView = mInflater.inflate(R.layout.fragment_random_contact_placeholder, mContainer, false);
+        TextView txtMessage = noAccessView.findViewById(R.id.txt_random_contact_placeholder);
+        txtMessage.setText(R.string.str_permission_required);
+        Button btnProvideAccess = noAccessView.findViewById(R.id.btnProvideAccess);
+        btnProvideAccess.setVisibility(View.GONE);
+        return noAccessView;
+    }
 
-        mContext = getActivity().getBaseContext();
+    @Override
+    public View setupViewWithPermission(Bundle savedInstanceState) {
+        View randomContactView;
+
         mContactManager = RandomContactManager.getInstance(mContext);
         mContactManagerProvider = (GlobalThemeChanger) getActivity();
 
-        initViews(layout);
+        if(mContactManager.getTotalContactCount() > 0){
+            randomContactView = mInflater.inflate(R.layout.fragment_random_contact, mContainer, false);
+            initViews(randomContactView);
 
-        setupRecyclerView(layout);
+            setupRecyclerView(randomContactView);
 
-        configureAnimations();
+            configureAnimations();
 
-        setupMenu();
+            setupMenu();
 
-        showInitialContact(savedInstanceState);
+            showInitialContact(savedInstanceState);
+        } else {
+            randomContactView = setupViewWithoutContacts();
+        }
 
-        return layout;
+        return randomContactView;
+    }
+
+    @Override
+    protected void onPermissionNotGranted(int requestCode, String[] permissions, int[] grantResults) {
+        Toast.makeText(mContext, R.string.str_permission_required, Toast.LENGTH_LONG).show();
     }
 
     private void showInitialContact(Bundle savedInstanceState) {
@@ -120,7 +140,7 @@ public class RandomContactFragment extends Fragment {
                         return true;
                     case R.id.action_view_contact:
                         Toast.makeText(getContext(), "Opening contact in Contacts app", Toast.LENGTH_SHORT).show();
-                        Utils.showContactInAddressBook(getActivity(), getCurrentContact().getId());
+                        Utils.showContactInAddressBook(requireActivity(), mCurrentContact.getId());
                         return true;
                 }
 
@@ -138,34 +158,34 @@ public class RandomContactFragment extends Fragment {
 
     private void setupRecyclerView(View layout) {
         // Create the Recycler Adapter
-        mRecyclerAdapter = new ContactRecyclerAdapter(Collections.EMPTY_LIST,
+        mRecyclerAdapter = new ContactRecyclerAdapter(Collections.<PhoneNumberVo>emptyList(),
                 R.layout.contact_info_item,
                 getActivity());
 
-        boolean whatsAppInstalled = Utils.isPackageInstalled(AppConstants.PACKAGE_NAME_WHATSAPP,
-                getActivity().getPackageManager());
-        mRecyclerAdapter.setEnableWhatsAppIntegration(whatsAppInstalled);
+        if(null != getActivity().getPackageManager()) {
+            boolean whatsAppInstalled = Utils.isPackageInstalled(AppConstants.PACKAGE_NAME_WHATSAPP, getActivity().getPackageManager());
+            mRecyclerAdapter.setEnableWhatsAppIntegration(whatsAppInstalled);
+        }
 
         // Find the RecyclerView and set some properties
-        RecyclerView recyclerView = (RecyclerView) layout.findViewById(android.R.id.list);
+        RecyclerView recyclerView = layout.findViewById(android.R.id.list);
         recyclerView.setHasFixedSize(true);
         recyclerView.setAdapter(mRecyclerAdapter);
         LinearLayoutManager linearLayoutManager = new LinearLayoutManager(mContext);
         recyclerView.setLayoutManager(linearLayoutManager);
         recyclerView.setItemAnimator(new DefaultItemAnimator());
-
     }
 
     private void initViews(View layout) {
         // Find some UI controls
-        mUserName = (TextView) layout.findViewById(R.id.userDisplayName);
-        mUserImage = (RoundedImageView) layout.findViewById(R.id.userProfileImage);
-        mUserContactedCount = (TextView) layout.findViewById(R.id.userContactedCount);
-        mLastContactedTime = (TextView) layout.findViewById(R.id.lastContactedTime);
-        mLastContactedLayout = (LinearLayout) layout.findViewById(R.id.lastContactedLayout);
-        mContactNowText = (TextView) layout.findViewById(R.id.contactNowText);
+        mUserName = layout.findViewById(R.id.userDisplayName);
+        mUserImage = layout.findViewById(R.id.userProfileImage);
+        mUserContactedCount = layout.findViewById(R.id.userContactedCount);
+        mLastContactedTime = layout.findViewById(R.id.lastContactedTime);
+        mLastContactedLayout = layout.findViewById(R.id.lastContactedLayout);
+        mContactNowText = layout.findViewById(R.id.contactNowText);
         mListContainer = layout.findViewById(R.id.listContainer);
-        mFragmentToolbar = (Toolbar) layout.findViewById(R.id.toolbar2);
+        mFragmentToolbar = layout.findViewById(R.id.toolbar2);
         mToolbarExtend = layout.findViewById(R.id.toolbarExtend);
 
         // Decode the default image and cache it
@@ -177,7 +197,7 @@ public class RandomContactFragment extends Fragment {
     }
 
     @Override
-    public void onSaveInstanceState(Bundle outState) {
+    public void onSaveInstanceState(@NonNull Bundle outState) {
         super.onSaveInstanceState(outState);
         if (null != mCurrentContact) {
             outState.putString(SAVED_CONTACT_ID, mCurrentContact.getId());
@@ -248,10 +268,6 @@ public class RandomContactFragment extends Fragment {
             Toast.makeText(mContext, getResources().getString(R.string.str_empty_contact_list), Toast.LENGTH_LONG)
                     .show();
         }
-    }
-
-    public ContactVo getCurrentContact() {
-        return mCurrentContact;
     }
 
 }