+++ randomContact/src/main/java/com/ae/apps/randomcontact/utils/AppConstants.java
@@ -11,4 +11,5 @@ public interface AppConstants {
     String PREF_KEY_NAV_DRAWER_INTRO_GIVEN = "pref_key_nav_drawer_intro";
 
     String PACKAGE_NAME_WHATSAPP = "com.whatsapp";
+
 }