+++ randomContact/src/main/java/com/ae/apps/randomcontact/utils/Utils.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2017 Midhun Harikumar
+ * Copyright 2017-2018 Midhun Harikumar
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,6 +27,10 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.telephony.PhoneNumberUtils;
 import android.text.format.DateUtils;
+import android.view.Gravity;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.FrameLayout;
 import android.widget.Toast;
 
 import java.text.ParseException;
@@ -80,6 +84,14 @@ public class Utils {
         }
 
         return friendlyDateString;
+    }
+
+    public static ViewGroup createParentLayout(final Context context){
+        FrameLayout layout = new FrameLayout(context);
+        FrameLayout.LayoutParams lp = new FrameLayout.LayoutParams(ViewGroup.LayoutParams.FILL_PARENT,
+                ViewGroup.LayoutParams.FILL_PARENT,Gravity.CENTER_HORIZONTAL| Gravity.CENTER_VERTICAL);
+        layout.setLayoutParams(lp);
+        return layout;
     }
 
     /**